pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "feedparser>=6.0.12",
    "google-api-python-client>=2.183.0",
    "google-genai>=1.39.1",
    "pandas>=2.3.3",
    "requests>=2.32.5",
    "sift-stack-py>=0.9.1",
    "streamlit>=1.50.0",
    "trafilatura>=2.0.0",
]
